[WA] Fix the sync issue between HVAC app and KitchSink app
Currently cel_apl used the default vehiclehal from Google, there's no
real vehicle hardware/system and it just used to simulate the vehiclehal,
when we set a property to vehicle, there's no feedback.
So when set a value on HVAC app or KitchSink app, there will be no notification
to the other app. It caused the value not match between these two applications.

Jira: OAM-65770
Test: Refer to the steps from JIRA
Signed-off-by: Lei,RayX <rayx.lei@intel.com>